Fix bld! (#163)
Ok, so this became much more complicated than I imagined.
bld is dependent on the bld script, which will include information about what class to execute.
Meaning by itself it can not be stored globally and run without the wrapper script.
To fix this gg will mimic the script, but without knowing the java file.
It will look for a java file and use the class name of this.
There can be multiple java files - to support this gg will first look in a bld file.
Best effort! :)