Update the download handling to use temporary directory for extraction and cleanup procedure (#181)
Ok, so the previous plan of snapshotting could work fine for a single user, buuuut for the integration tests it was madness.
This fix uses temp dirs. I contemplated using temp dirs long ago but decided against it, but now it's on again.
But, there is a difference - actual download still goes to download because:
  * Can continue (_can_ continue lol) on download pause/stop/network issue/whatever
  * Can re-extract if extract fails, or is canceled of whatever
Then output of extract, both first and second, is in temp.