Add `.jar` file support for custom commands and GitHub executors
Includes handling of `.jar` files as input URLs for downloads, dependencies, and binary patterns. Enhanced argument customization and introduced a utility to identify `.jar` files in installation paths.
Note that this won't append java as dep if jar is in release files, because it calculates deps beforehand.
I don't know how to get around this, besides using the api to query for files before completing the resolve, but I feel that's simply "too much".